Add dcline handling for call from most.
The args argument contains the mpc and vs struct.
We distinguish between call from most or opf by checking the om model class.
If om is opt_model class object, the call comes from most.
If the class object is a opf_model the call comes from opf.

The strcomp is necessary because isa does not distinguish between
class and its respective sub-classes.